Update new projects document
Some of the things originally in this document probably aren't as relevant to such a small project as the Open Data Cube, for example demonstration of an active user community - often projects in this organisation will only (initially) have users who are the developers of the project. Some of these changes are more suggestions and open to discussion.